// Cambio a la modularizacion de las importaciones //
Old  ---> 'src/users/usersService'

New  ---> '../users/usersService'

+++ src/users/users.controller.ts
@@ -13,12 +13,12 @@ import {
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-users.dto';
-import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { JwtAuthGuard } from '../auth/jwt-auth.guard';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 import { FileInterceptor } from '@nestjs/platform-express';
-import { multerConfig } from 'src/file/multer.config';
-import { FileService } from 'src/file/file.service';
-import { GetUser } from 'src/auth/decorator/get-user.decorator';
+import { multerConfig } from '../file/multer.config';
+import { FileService } from '../file/file.service';
+import { GetUser } from '../auth/decorator/get-user.decorator';
 
 @ApiBearerAuth()
 @ApiTags('Users')
@@ -27,23 +27,25 @@ export class UsersController {
   constructor(private usersService: UsersService,
     private fileService: FileService) {}
 
+  @UseGuards(JwtAuthGuard)
   @Get()
   getAllUsers() {
     return this.usersService.findAll();
   }
 
-  // @Post()
-  // async createUser(@Body() newUser: CreateUserDto) {
-  //   return this.usersService.createUser(newUser);
-  // }
-
+  @UseGuards(JwtAuthGuard)
   @Get(':id')
   async findById(@Param('id') id: string) {
     return this.usersService.findById(id);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  async deleteById(@Param('id') id: string) {
+  async deleteById(
+  @GetUser() user: any,
+    @Param('id') id: string,
+  ) {
+    if (user.id !== id) return { resp: 'Forbidden resource', status: HttpStatus.FORBIDDEN };
     return this.usersService.delete(id);
   }
 
@@ -52,7 +54,12 @@ export class UsersController {
   @UseInterceptors(
     FileInterceptor('profilePic', multerConfig),
   )
-  async updateUser(@GetUser() user: any, @Param('id') id: string, @Body() body: CreateUserDto, @UploadedFile() profilePic?: Express.Multer.File ) {
+  async updateUser(
+  @GetUser() user: any,
+    @Param('id') id: string,
+    @Body() body: CreateUserDto,
+    @UploadedFile() profilePic?: Express.Multer.File,
+  ) {
     if (user.id !== id) return { resp: 'Forbidden resource', status: HttpStatus.FORBIDDEN };
     if (profilePic) {
       const url = await this.fileService.createFiles(profilePic);
